move tutorial into sphinx resources

+++ doc_source/documentDASH.m
@@ -1,4 +1,4 @@
-function[] = documentDASH
+function[] = documentDASH(rebuildReference)
 %% documentDASH  Builds the complete reference guide for the DASH toolbox
 % ----------
 %   documentDASH
@@ -12,7 +12,14 @@ function[] = documentDASH
 %   markup (using the config and index files in doc_source/sphinx/config)
 %   to generate the final HTML reference set. The final set is located in
 %   doc_source/sphinx/build/html.
+%
+%   documentDASH(rebuildReference)
+%   Indicate whether to rebuild the RST for the reference guide
 % ----------
+%   Inputs:
+%       rebuildReference (scalar logical): True if the command should
+%           rebuild the RST pages for the reference guide. False to skip
+%           building the reference guide RSTs.
 
 
 %%%%% Parameters
@@ -20,7 +27,7 @@ function[] = documentDASH
 % File locations (relative to the location of this function)
 examplesLocation = "examples";
 rstDestination = "sphinx/source";
-sphinxConfig = "sphinx/config";
+sphinxResources = "sphinx/resources";
 htmlBuild = "build/html";
 doctrees = "build/doctrees";
 
@@ -38,9 +45,12 @@ contents = ["gridMetadata",      "class"
             "dash",              "package"];
 %%%%%
 
-%% Prepare source
-% This section prepares the source folder to receive the RST files
-fprintf('Preparing source\n');
+
+%% Setup
+
+if ~exist('rebuildReference', 'var')
+    rebuildReference = true;
+end
 
 % Get the location of this function
 fullpath = mfilename('fullpath');
@@ -48,45 +58,55 @@ folders = split(fullpath, filesep);
 folders = folders(1:end-1);
 docSource = strjoin(folders, filesep);
 
+% Locate the source folder
+source = fullfile(docSource, rstDestination);
+
 % Move to the location of "buildDocs", but return to current location when finished
 currentDir = pwd;
 goback = onCleanup( @()cd(currentDir) );
 cd(docSource);
 
-% Clear the source .rst folder and move to it
-source = fullfile(docSource, rstDestination);
-if isfolder(source)
-    rmdir(source, 's');
+
+%% Prepare source
+% This section prepares the source folder to receive the RST files
+
+if rebuildReference
+    fprintf('Preparing source\n');
+    
+    % Clear the source .rst folder and move to it
+    if isfolder(source)
+        rmdir(source, 's');
+    end
+    mkdir(source);
+    cd(source);
+    
+    %% Build RST
+    % This section builds the RST files for each item in DASH
+    fprintf('Building reST files\n');
+    
+    % Document each item
+    examplesRoot = fullfile(docSource, examplesLocation);
+    nContents = size(contents, 1);
+    for c = 1:nContents
+        item = contents(c,1);
+        type = contents(c,2);
+        examples = strcat(examplesRoot, filesep, item);
+        document.(type)(item, examples);
+    end
+    
+    % Write the documentation for PSM.template
+    writePSMTemplate;
+    
+    % Fix the links to inherited methods
+    fixInheritanceLinks;
 end
-mkdir(source);
-cd(source);
-
-%% Build RST
-% This section builds the RST files for each item in DASH
-fprintf('Building reST files\n');
-
-% Document each item
-examplesRoot = fullfile(docSource, examplesLocation);
-nContents = size(contents, 1);
-for c = 1:nContents
-    item = contents(c,1);
-    type = contents(c,2);
-    examples = strcat(examplesRoot, filesep, item);
-    document.(type)(item, examples);
-end
-
-% Write the documentation for PSM.template
-writePSMTemplate;
-
-% Fix the links to inherited methods
-fixInheritanceLinks;
 
 %% Run sphinx
 % This section runs sphinx in order to generate HTML from the RST
 fprintf('Preparing sphinx\n');
 
 % Add the sphinx conf.py, index.rst, and _static folder to the source
-conf = fullfile(docSource, sphinxConfig);
+conf = fullfile(docSource, sphinxResources);
 copyfile(conf, source);
 
 % Clear the HTML build
